refactor(android): v3.0 restructure text span
