+++ renderer/native/android/src/main/java/com/tencent/renderer/component/text/TextLetterSpacingSpan.java
@@ -1,0 +1,46 @@
+/* Tencent is pleased to support the open source community by making Hippy available.
+ * Copyright (C) 2018 THL A29 Limited, a Tencent company. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.tencent.renderer.component.text;
+
+import android.annotation.TargetApi;
+import android.os.Build;
+import android.text.TextPaint;
+import android.text.style.MetricAffectingSpan;
+import androidx.annotation.NonNull;
+
+@TargetApi(Build.VERSION_CODES.LOLLIPOP)
+public class TextLetterSpacingSpan extends MetricAffectingSpan {
+
+    private final float mSpace;
+
+    public TextLetterSpacingSpan(float space) {
+        mSpace = space;
+    }
+
+    @Override
+    public void updateMeasureState(@NonNull TextPaint textPaint) {
+        if (!Float.isNaN(mSpace)) {
+            textPaint.setLetterSpacing(mSpace / textPaint.getTextSize());
+        }
+    }
+
+    @Override
+    public void updateDrawState(@NonNull TextPaint textPaint) {
+        if (!Float.isNaN(mSpace)) {
+            textPaint.setLetterSpacing(mSpace / textPaint.getTextSize());
+        }
+    }
+}
